requirements.txt
pandas==2.0.3
plotly==5.18.0
python-dotenv==1.0.0
scikit_learn==1.2.2
scikit_learn==1.3.0
spotipy==2.23.0
tqdm==4.66.1
